xTile Collision
Hi, Faham
I have cleaned up the code and seperated it into two seperate classes.
Player and Game1. I have also added code in the player class to make the
player animated. My problem is the collisions still? It is colliding but
it is colliding in the wrong location. It seems to me to be a logic
error but I can't seem to find it.